Switch batch configuration to contain just one job with two steps
That is all we really need
